move locate trait to AstResolver
